Move tests and fix configurations
