+++ .github/workflows/config-file-deps-bump.yml
@@ -11,4 +11,4 @@ jobs:
     uses: beeware/.github/.github/workflows/dep-version-bump.yml@main
     secrets: inherit
     with:
-      subdirectory: . core dummy android cocoa demo gtk iOS testbed textual toga web winforms
+      subdirectory: . core dummy android cocoa demo gtk iOS testbed textual toga travertino web winforms
