+++ .github/workflows/publish.yml
@@ -27,6 +27,7 @@ jobs:
         - "toga_iOS"
         - "toga_web"
         - "toga_winforms"
+        - "travertino"
     steps:
       - name: Get packages
         uses: dsaltares/fetch-gh-release-asset@1.1.2
