+++ tox.ini
@@ -36,7 +36,7 @@ allowlist_externals =
 commands =
     # TOGA_INSTALL_COMMAND is set to a bash command by the CI workflow
     # Install as editable so parallel runs don't clobber the build directory for each other
-    {env:TOGA_INSTALL_COMMAND:python -m pip install {tox_root}{/}core[dev] {tox_root}{/}dummy}
+    {env:TOGA_INSTALL_COMMAND:python -m pip install {tox_root}{/}core[dev] {tox_root}{/}dummy {tox_root}{/}travertino}
     !cov: python -m pytest {posargs:-vv --color yes}
     cov : python -m coverage run -m pytest {posargs:-vv --color yes}
