--- travertino/tox.ini
@@ -1,41 +1,0 @@
-# Flake8 doesn't believe in pyproject.toml, so we put the configuration here.
-[flake8]
-exclude=
-    local/*,\
-    docs/*,\
-    build/*,\
-    dist/*,\
-    .tox/*, \
-    venv*
-max-complexity = 25
-max-line-length = 119
-ignore =
-    # line break occurred before a binary operator
-    W503,
-
-[tox]
-envlist = towncrier-check,pre-commit,py{39,310,311,312,313,314}
-skip_missing_interpreters = true
-
-[testenv:pre-commit]
-package = wheel
-wheel_build_env = .pkg
-extras = dev
-commands = pre-commit run --all-files --show-diff-on-failure --color=always
-
-[testenv:py{,39,310,311,312,313,314}]
-package = wheel
-wheel_build_env = .pkg
-depends =
-    towncrier-check
-    pre-commit
-extras = dev
-commands = python -m pytest {posargs:-vv --color yes}
-
-[testenv:towncrier{,-check}]
-skip_install = True
-deps =
-    towncrier==24.8.0
-commands =
-    check  : python -m towncrier.check --compare-with origin/main
-    !check : python -m towncrier {posargs}
